Fixed corrupt form
Chirag: DO NOT MODIFY/REMOVE the GEN-BEGIN,END,FIRST,LAST comments;
They are put there for a reason! When they are removed, the form gets
corrupted and the swing builder gets broken and the design tab is
essentially disabled in the process. initComponents is not meant to be
editable. If you need to change something in there, create a seperate
method (i.e. myInitComponents) and make your changes there. Goodjob
fixing the dropdowns though.